New preferences and auto-translate.
The extension can now always translate or never translate on a language by
language basis. The preference for each language can be set in the new
preferences page or whenever you encounter a page in that language.

The native language to translate into is now chosen in the first instance by
detecting the browser locale; this can be changed in the preferences as well.